# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Freelancers",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        },
        {
            "url": "http://api-freelancer.otterscompany.com"
        }
    ],
    "paths": {
...
        "/api/v1/settings": {
            "get": {
                "tags": [
                    "Settings"
                ],
                "summary": "Get settings",
                "description": "Returns all settings",
                "operationId": "indexSetting",
                "responses": {
                    "200": {
                        "description": "Show all settings.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Not authenticated."
                    },
                    "403": {
                        "description": "Forbidden."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Settings"
                ],
                "summary": "Update settings",
                "description": "Update settings",
                "operationId": "updateSetting",
                "parameters": [
                    {
                        "name": "stripe_public",
                        "in": "query",
                        "description": "Stripe public key",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "stripe_secret",
                        "in": "query",
                        "description": "Stripe secret key",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "paypal_public",
                        "in": "query",
                        "description": "Paypal public key",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "paypal_secret",
                        "in": "query",
                        "description": "Paypal secret key",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update settings.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Not authenticated."
                    },
                    "403": {
                        "description": "Forbidden."
                    },
                    "422": {
                        "description": "Data not valid."
                    },
                    "500": {
                        "description": "An error occurred during the process."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}